move date and time formatting functions to utils/dateAndTime.ts file
import back into Calendar.tsx
