+++ app/components/calendar/Calendar.tsx
@@ -17,6 +17,7 @@ import {
   SelectedRangeType,
   TimeBlockType,
 } from "@/types";
+import { getLocalDateString, getSlotMinTime } from "@/utils/dateAndTime";
 
 export default function CalendarView() {
   const [events, setEvents] = useState<CalendarEventType[]>([]);
@@ -110,17 +111,6 @@ export default function CalendarView() {
     }
   }, []);
 
-  function getLocalDateString(): string {
-    const today = new Date();
-    return (
-      today.getFullYear() +
-      "-" +
-      String(today.getMonth() + 1).padStart(2, "0") +
-      "-" +
-      String(today.getDate()).padStart(2, "0")
-    );
-  }
-
   const handleEventClick = async (clickInfo: EventClickArg) => {
     const clientId = localStorage.getItem("clientId");
     if (clickInfo.event.extendedProps.clientId === clientId) {
@@ -128,18 +118,6 @@ export default function CalendarView() {
       setEditValue(clickInfo.event.title);
     }
   };
-
-  function getSlotMinTime(viewDate: Date): string {
-    const now = new Date();
-    const isToday = now.toDateString() === viewDate.toDateString();
-
-    if (isToday) {
-      const hours = now.getHours().toString().padStart(2, "0");
-      return `${hours}:00:00`; // rounded down to the hour
-    }
-
-    return "00:00:00"; // full day for other dates
-  }
 
   return (
     <>
